Remove redux
