+++ package.json
@@ -42,12 +42,8 @@
     "react-document-title": "^2.0.2",
     "react-dom": "^15.5.4",
     "react-json-view": "^1.8.2",
-    "react-redux": "^5.0.3",
     "react-router": "3.0.0",
-    "react-router-redux": "^4.0.8",
     "react-scrollable-anchor": "^0.4.2",
-    "redux": "^3.6.0",
-    "redux-thunk": "^2.2.0",
     "superagent": "^3.5.2"
   },
   "devDependencies": {
