--- src/store/definition/constants.js
@@ -1,6 +1,0 @@
-export const ActionType = {
-  FETCH_DEFINITION_SUCCESS: 'FETCH_DEFINITION_SUCCESS',
-  FETCH_DEFINITION_FAILURE: 'FETCH_DEFINITION_FAILURE',
-  PARSE_DEFINITION_SUCCESS: 'PARSE_DEFINITION_SUCCESS',
-  PARSE_DEFINITION_FAILURE: 'PARSE_DEFINITION_FAILURE'
-}
